pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "video-summarize"
version = "0.1.0"
description = "Video summarization with BiLSTM (supervised)"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "numpy>=1.24.0",
    "PyYAML>=6.0",
    "opencv-python>=4.8.0",
    "scipy>=1.10.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
